refactor(auth): 重构用户认证逻辑
- 修改 API 基础 URL，移除末尾斜杠
- 更新用户登录和认证相关代码
- 新增 auth 工具函数，用于处理 token
- 调整 store 结构，将 user store 移至 composables 文件夹
