+++ .env.dev
@@ -2,4 +2,4 @@
 VITE_ENV = 'development'
 
 # 基础配置
-VITE_BASE_API = 'http://111.229.29.214:8080/'
+VITE_BASE_API = 'http://111.229.29.214:8080'
